+++ index.html
@@ -8,7 +8,6 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
 </head>
 
 <body>
@@ -233,6 +232,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
